refactor(daemon): move wardnetd-services inline tests to tests/ layout
Migrate the nine inline test modules to the mandated layout
(.agents/testing.md):

- dns/log_sink.rs, dns/dhcp_lan_runner.rs (+ its dhcp_lan_runner/tests.rs
  sub-file), push/sender.rs and update/service.rs are merged into their
  pre-existing sibling suites (dns/tests/log_sink.rs,
  dns/tests/dhcp_lan_runner.rs, push/tests.rs, update/tests/service.rs)
  with no test cases dropped and no name collisions.
- dns/service.rs's forwarder_selection_tests + forwarder_service_tests
  move to a new dns/tests/service.rs.
- tunnel/service.rs's summarize_latency_tests merge into the existing
  tunnel suite in tunnel/tests/tunnel.rs.
- subnet.rs and entitlement.rs move to src/tests/{subnet,entitlement}.rs
  alongside the other crate-root module tests.
- rule_request/tests/mod.rs is split into rule_request/tests/service.rs,
  matching zone_exception/zone_enforcement/network_zone.

resolve_forwarder_selection, is_newer, build_request and the
dhcp_lan_runner internals the tests exercise are widened to pub(crate);
test bodies are unchanged.

Part of #847.
