+++ source/daemon/crates/wardnetd-services/src/dns/tests/log_sink.rs
@@ -120,3 +120,102 @@ async fn new_with_stats_records_by_domain_only_for_blocked() {
         "domain label must include the blocked domain"
     );
 }
+
+fn make_meter() -> Meter {
+    Meter::new(StatsBuffer::new())
+}
+
+fn make_row(domain: &str, device_id: Option<&str>) -> QueryLogRow {
+    QueryLogRow {
+        timestamp: "2026-06-12T00:00:00Z".to_owned(),
+        client_ip: "192.168.1.1".to_owned(),
+        domain: domain.to_owned(),
+        query_type: "A".to_owned(),
+        result: "forwarded".to_owned(),
+        upstream: None,
+        latency_ms: 1.0,
+        device_id: device_id.map(str::to_owned),
+    }
+}
+
+/// A row with `device_id = Some(...)` must appear on both `capture_rx`
+/// and `persist_rx`.
+#[test]
+fn capture_forwarded_when_device_id_set() {
+    let meter = make_meter();
+    let (sink, mut channels) = DnsLogSink::new_with_stats(&meter);
+
+    sink.record(make_row("example.com", Some("dev-1")));
+
+    let captured = channels
+        .capture_rx
+        .try_recv()
+        .expect("expected row on capture_rx");
+    assert_eq!(captured.domain, "example.com");
+
+    let persisted = channels
+        .persist_rx
+        .try_recv()
+        .expect("expected row on persist_rx");
+    assert_eq!(persisted.domain, "example.com");
+}
+
+/// A row with `device_id = None` must NOT be forwarded to `capture_rx`.
+#[test]
+fn capture_skipped_without_device_id() {
+    let meter = make_meter();
+    let (sink, mut channels) = DnsLogSink::new_with_stats(&meter);
+
+    sink.record(make_row("example.com", None));
+
+    assert_eq!(
+        channels.capture_rx.try_recv().unwrap_err(),
+        tokio::sync::mpsc::error::TryRecvError::Empty,
+    );
+
+    // Persist channel should still receive the row.
+    assert!(channels.persist_rx.try_recv().is_ok());
+}
+
+/// `take_capture_dropped()` starts at 0 and resets to 0 after being read.
+#[test]
+fn capture_dropped_counter_starts_zero_and_resets() {
+    let meter = make_meter();
+    let (sink, _channels) = DnsLogSink::new_with_stats(&meter);
+
+    assert_eq!(sink.take_capture_dropped(), 0);
+    // A second call must also return 0 (counter was reset by the first call).
+    assert_eq!(sink.take_capture_dropped(), 0);
+}
+
+/// `record_dns_stats` records the `by_domain` counter for blocked queries.
+#[test]
+fn blocked_outcome_records_by_domain_stat() {
+    let meter = make_meter();
+    let (sink, _channels) = DnsLogSink::new_with_stats(&meter);
+
+    let mut row = make_row("blocked-ads.tracker.io", Some("dev-1"));
+    row.result = "blocked".to_owned();
+    sink.record(row);
+
+    // The only assertion needed for coverage is that we don't panic;
+    // the by_domain counter is an internal implementation detail.
+}
+
+/// When the persist channel is full, `dropped_entries` is incremented.
+#[test]
+fn persist_full_increments_dropped_counter() {
+    // Capacity-1 persist channel: first send fills it, second is dropped.
+    let (sink, mut persist_rx) = DnsLogSink::with_capacities(1, 256);
+
+    sink.record(make_row("first.example.com", None));
+    sink.record(make_row("second.example.com", None));
+
+    // Exactly one row should be queued, one should have been dropped.
+    assert_eq!(persist_rx.try_recv().unwrap().domain, "first.example.com");
+    assert!(
+        persist_rx.try_recv().is_err(),
+        "second row should have been dropped"
+    );
+    assert_eq!(sink.take_dropped(), 1);
+}
