+++ source/daemon/crates/wardnetd-services/src/dns/tests/mod.rs
@@ -5,3 +5,4 @@ mod dhcp_lan_runner;
 mod filter_parser;
 mod log_sink;
 mod query_log_runner;
+mod service;
