+++ source/daemon/crates/wardnetd-services/src/push/tests.rs
@@ -27,7 +27,7 @@ use wardnetd_data::repository::{
 use wardnetd_data::secret_store::SecretStore;
 
 use crate::auth_context;
-use crate::push::sender::{PushTarget, SendOutcome, VapidKey, WebPushSender};
+use crate::push::sender::{PushTarget, ReqwestWebPushSender, SendOutcome, VapidKey, WebPushSender};
 use crate::push::{PushService, PushServiceImpl};
 
 /// The daemon-seeded Guest zone (default-for-new); used as a valid `zone_id` FK
@@ -997,4 +997,175 @@ async fn device_keyed_payload_omits_url_but_keeps_kind_and_subject() {
     // Absent keys are omitted, not serialized as null.
     assert!(payload["data"].get("url").is_none());
     assert_eq!(payload["data"]["subject_id"], device_id.to_string());
+}
+
+// ── ReqwestWebPushSender (wire format + outcome classification) ─────────
+
+use base64::Engine as _;
+use base64::engine::general_purpose::URL_SAFE_NO_PAD;
+
+/// A valid subscriber target: a real P-256 point + 16-byte auth secret.
+fn valid_target(endpoint: &str) -> (String, String, String) {
+    let p256dh = VapidKey::generate().public_key_base64url();
+    let auth = URL_SAFE_NO_PAD.encode([9u8; 16]);
+    (endpoint.to_owned(), p256dh, auth)
+}
+
+async fn send_to(status: u16) -> SendOutcome {
+    use wiremock::matchers::method;
+    use wiremock::{Mock, MockServer, ResponseTemplate};
+
+    let server = MockServer::start().await;
+    Mock::given(method("POST"))
+        .respond_with(ResponseTemplate::new(status))
+        .mount(&server)
+        .await;
+
+    let sender = ReqwestWebPushSender::new(reqwest::Client::new(), "mailto:a@b.c".to_owned());
+    let vapid = VapidKey::generate();
+    let (endpoint, p256dh, auth) = valid_target(&server.uri());
+    sender
+        .send(
+            &vapid,
+            PushTarget {
+                endpoint: &endpoint,
+                p256dh: &p256dh,
+                auth: &auth,
+            },
+            b"{\"title\":\"hi\"}".to_vec(),
+        )
+        .await
+}
+
+#[tokio::test]
+async fn send_classifies_http_responses() {
+    assert_eq!(send_to(201).await, SendOutcome::Delivered);
+    assert_eq!(send_to(200).await, SendOutcome::Delivered);
+    assert_eq!(send_to(410).await, SendOutcome::Gone);
+    assert_eq!(send_to(404).await, SendOutcome::Gone);
+    assert_eq!(send_to(500).await, SendOutcome::TransientFailure);
+    assert_eq!(send_to(429).await, SendOutcome::TransientFailure);
+}
+
+#[tokio::test]
+async fn send_maps_unbuildable_subscription_to_gone() {
+    // A malformed p256dh can never encrypt -> pruned as Gone.
+    let sender = ReqwestWebPushSender::new(reqwest::Client::new(), "mailto:a@b.c".to_owned());
+    let vapid = VapidKey::generate();
+    let outcome = sender
+        .send(
+            &vapid,
+            PushTarget {
+                endpoint: "https://push.example.com/x",
+                p256dh: "not-a-valid-point",
+                auth: "YXV0aA",
+            },
+            b"{}".to_vec(),
+        )
+        .await;
+    assert_eq!(outcome, SendOutcome::Gone);
+}
+
+#[tokio::test]
+async fn send_maps_network_error_to_transient() {
+    // Nothing is listening on this port -> connection refused.
+    let sender = ReqwestWebPushSender::new(reqwest::Client::new(), "mailto:a@b.c".to_owned());
+    let vapid = VapidKey::generate();
+    let (endpoint, p256dh, auth) = valid_target("http://127.0.0.1:1/push");
+    let outcome = sender
+        .send(
+            &vapid,
+            PushTarget {
+                endpoint: &endpoint,
+                p256dh: &p256dh,
+                auth: &auth,
+            },
+            b"{}".to_vec(),
+        )
+        .await;
+    assert_eq!(outcome, SendOutcome::TransientFailure);
+}
+
+#[test]
+fn from_bytes_rejects_wrong_length_without_panicking() {
+    let Err(err) = VapidKey::from_bytes(b"too-short") else {
+        panic!("expected a length error");
+    };
+    assert!(err.to_string().contains("32 bytes"), "got {err}");
+}
+
+#[test]
+fn vapid_public_key_is_stable_across_reload() {
+    let key = VapidKey::generate();
+    let public = key.public_key_base64url();
+    let bytes = key.to_bytes();
+
+    let reloaded = VapidKey::from_bytes(&bytes).unwrap();
+    assert_eq!(reloaded.public_key_base64url(), public);
+    // Uncompressed P-256 point is 65 bytes -> 87 base64url chars unpadded.
+    assert_eq!(public.len(), 87);
+}
+
+#[test]
+fn build_request_sets_web_push_headers() {
+    // A syntactically valid subscription (keys are real base64url of the
+    // right lengths) must produce an aes128gcm-encoded POST with a VAPID
+    // Authorization header.
+    let vapid = VapidKey::generate();
+    // 65-byte uncompressed P-256 point for a throwaway key = valid p256dh.
+    let subscriber = VapidKey::generate();
+    let p256dh = subscriber.public_key_base64url();
+    let auth = base64::engine::general_purpose::URL_SAFE_NO_PAD.encode([7u8; 16]);
+
+    let sender = ReqwestWebPushSender::new(reqwest::Client::new(), "mailto:a@b.c".to_owned());
+    let request = sender
+        .build_request(
+            &vapid,
+            &PushTarget {
+                endpoint: "https://push.example.com/xyz",
+                p256dh: &p256dh,
+                auth: &auth,
+            },
+            b"{\"title\":\"hi\"}".to_vec(),
+        )
+        .expect("valid subscription should build");
+
+    assert_eq!(request.method(), http::Method::POST);
+    assert_eq!(
+        request
+            .headers()
+            .get(http::header::CONTENT_ENCODING)
+            .unwrap(),
+        "aes128gcm"
+    );
+    let auth_header = request
+        .headers()
+        .get(http::header::AUTHORIZATION)
+        .unwrap()
+        .to_str()
+        .unwrap();
+    assert!(auth_header.starts_with("vapid t="), "got {auth_header}");
+}
+
+#[test]
+fn build_request_rejects_wrong_length_auth() {
+    let vapid = VapidKey::generate();
+    let subscriber = VapidKey::generate();
+    let p256dh = subscriber.public_key_base64url();
+    // 15 bytes instead of 16.
+    let auth = base64::engine::general_purpose::URL_SAFE_NO_PAD.encode([7u8; 15]);
+
+    let sender = ReqwestWebPushSender::new(reqwest::Client::new(), "mailto:a@b.c".to_owned());
+    let err = sender
+        .build_request(
+            &vapid,
+            &PushTarget {
+                endpoint: "https://push.example.com/xyz",
+                p256dh: &p256dh,
+                auth: &auth,
+            },
+            b"{}".to_vec(),
+        )
+        .unwrap_err();
+    assert!(err.to_string().contains("16 bytes"), "got {err}");
 }
